Change behaviour of custom adjustment to a more expected one.
Before it multiplied the default one by the custom one.
Now it completely discards the default one.
It allows to, for example, set some levels to be ineffective
without having to calculate the coefficent which will bring them
back to flow 1.

A fixture to test this behaviour was added and some support code
to test on fixtures too.